#3. Create Map View on app home page
