+++ .watchmanconfig
