+++ android/settings.gradle
@@ -1,0 +1,5 @@
+rootProject.name = 'CuriousEdinburgh'
+include ':react-native-maps'
+project(':react-native-maps').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-maps/android')
+
+include ':app'
