+++ index.android.js
@@ -1,0 +1,4 @@
+import { AppRegistry } from 'react-native';
+import CuriousEdinburgh from './src/components/CuriousEdinburgh';
+
+AppRegistry.registerComponent('CuriousEdinburgh', () => CuriousEdinburgh);
