+++ index.ios.js
@@ -1,0 +1,5 @@
+
+import { AppRegistry } from 'react-native';
+import CuriousEdinburgh from './src/components/CuriousEdinburgh';
+
+AppRegistry.registerComponent('CuriousEdinburgh', () => CuriousEdinburgh);
