+++ ios/CuriousEdinburgh.xcodeproj/xcshareddata/xcschemes/CuriousEdinburgh.xcscheme
