+++ ios/CuriousEdinburgh/AppDelegate.h
