+++ ios/CuriousEdinburgh/AppDelegate.m
