+++ ios/CuriousEdinburgh/Images.xcassets/AppIcon.appiconset/Contents.json
