+++ ios/CuriousEdinburgh/Info.plist
