+++ ios/CuriousEdinburghTests/Info.plist
