+++ src/components/CuriousEdinburgh.js
@@ -1,0 +1,10 @@
+import React, { Component } from 'react';
+import Home from './Home';
+
+export default class CuriousEdinburgh extends Component {
+  render() {
+    return (
+      <Home />
+    );
+  }
+}
